requirements.txt
alembic==1.4.3
aniso8601==8.0.0
asgiref==3.4.1
attrs==20.3.0
Automat==0.8.0
bcrypt==3.1.7
beautifulsoup4==4.10.0
blinker==1.4
boto3==1.13.9
botocore==1.16.26
cachetools==4.2.4
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cmarkgfm==0.4.2
configobj==5.0.6
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
dataset==1.3.1
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.4.0
dj-database-url==0.5.0
Django==3.1.13
django-allauth==0.45.0
django-cors-headers==3.10.0
django-filter==21.1
django-heroku==0.3.1
django-model-utils==4.0.0
django-oauth-toolkit==1.5.0
django-phonenumber-field==5.2.0
django-rest-auth==0.9.5
django-rest-knox==4.1.0
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-x509==0.9.4
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
docutils==0.15.2
drf-yasg==1.20.0
entrypoints==0.3
et-xmlfile==1.1.0
Flask==1.1.2
Flask-Caching==1.8.0
flask-marshmallow==0.10.1
Flask-Migrate==2.5.3
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.3
gevent==20.9.0
google-api-core==2.2.2
google-api-python-client==2.29.0
google-auth==2.3.3
google-auth-httplib2==0.1.0
googleapis-common-protos==1.53.0
greenlet==0.4.17
gunicorn==20.0.4
htmlmin==0.1.12
httplib2==0.20.2
hyperlink==19.0.0
idna==2.10
image==1.5.33
importlib-metadata==1.5.0
incremental==16.10.1
inflection==0.5.1
itsdangerous==1.1.0
itypes==1.2.0
Jinja2==2.11.3
jmespath==0.10.0
jsonfield==3.1.0
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lxml==4.6.4
Mako==1.1.3
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==2.20.2
marshmallow-sqlalchemy==0.17.0
maxminddb==1.5.4
more-itertools==4.2.0
netifaces==0.10.4
oauth2client==4.1.3
oauthlib==3.1.0
openapi-codec==1.3.2
openpyxl==3.0.9
openwisp-utils==0.7.5
packaging==21.0
passlib==1.7.2
pexpect==4.6.0
Pillow==8.4.0
protobuf==3.19.1
psycopg2==2.9.2
psycopg2-binary==2.9.1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pybluemonday==0.0.8
pycparser==2.20
pydantic==1.6.2
PyDrive==1.3.1
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==2.3.0
pymacaroons==0.13.0
PyMySQL==0.9.3
PyNaCl==1.3.0
pyOpenSSL==20.0.1
pyparsing==3.0.5
pyrsistent==0.17.3
pyserial==3.4
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
python-geoacumen-city==0.0.1
python3-openid==3.2.0
pytz==2020.4
PyYAML==5.3.1
qrcode==7.3.1
redis==3.5.2
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
s3transfer==0.3.3
SecretStorage==2.3.1
service-identity==18.1.0
simplejson==3.16.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
soupsieve==2.3.1
SQLAlchemy==1.3.17
SQLAlchemy-Utils==0.36.6
sqlparse==0.4.2
ssh-import-id==5.10
stripe==2.61.0
swagger-spec-validator==2.7.3
swapper==1.1.2.post1
systemd-python==234
tenacity==6.2.0
Twisted==18.9.0
uritemplate==4.1.1
urllib3==1.25.11
wadllib==1.3.3
Werkzeug==1.0.1
whitenoise==6.0.0
wrapt==1.13.2
WTForms==2.3.1
zipp==1.0.0
zope.event==4.5.0
zope.interface==5.2.0